requirements.txt
anyio==3.7.0
click==8.1.3
Deprecated==1.2.14
fastapi==0.95.2
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
limits==3.5.0
packaging==23.1
pydantic==1.10.8
python-dotenv==1.0.0
redis==4.5.5
slowapi==0.1.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
wrapt==1.15.0
